* compile in VS2005

+++ src/lastactivity.cpp
@@ -65,7 +65,7 @@ namespace gloox
         t->addAttribute( "id", stanza->id() );
         t->addAttribute( "to", stanza->from().full() );
         Tag *q = new Tag( t, "query" );
-        q->addAttribute( "seconds", now - m_active );
+        q->addAttribute( "seconds", (long)( now - m_active ) );
         q->addAttribute( "xmlns", XMLNS_LAST );
 
         m_parent->send( t );
